requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-cors>=4.0
PyJWT>=2.8
docker>=7.0
pyyaml>=6.0
jinja2>=3.1
requests>=2.31
psutil>=5.9
APScheduler>=3.10
gunicorn>=21.0